# util/encoding: Replace byte iteration with bytes.IndexByte in findDecimalTerminator