This addresses a TODO that was waiting on Go 1.7.

The only benchmark this seemed to move was `PeekLengthDecimal`, which
got a nice 14% speed bump:

```
name                 old time/op    new time/op    delta
PeekLengthDecimal-4    48.4ns ± 1%    41.6ns ± 1%  -14.09%  (p=0.000 n=10+9)

name                 old alloc/op   new alloc/op   delta
PeekLengthDecimal-4    0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)

name                 old allocs/op  new allocs/op  delta
PeekLengthDecimal-4     0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/10035)
<!-- Reviewable:end -->